Add basic setup.py
This will allow the project to be easily installed in other projects,
even if it isn't published to pypi.

For example, in requirements.txt one could add:

-e
git://github.com/borgbase/borgbase-api-client.git#egg=borgbase-api-client

I chose the requests version >= 2.21.0 because that is what is currently available
in debian stable.